Use CONTENT_DELAYIFSLOW.
In Total Commander, only one ContentGetValue running at the same time (either
from the main or background thread), so there is no need to protect the cache
for multithreaded access.

In Double Commander, CONTENT_DELAYIFSLOW is currently not implemented. When it
will, the code needs to be properly tested, although the behavior is most
likely to be compatible with Total Commander.