Require FPC 3.1.1
The current code in fact doesn't compile with FPC 2.6.0:

* EArgumentException (used by TLineReader) was added in 2.6.4
* PSizeUint (used in brotli.pas) was added in 3.1.1:
  https://gitlab.com/freepascal.org/fpc/source/-/issues/28940

It's easy to add these types for compatibility with 2.6.0, but it
probably doesn't make sense to bother. The only downside is that
Ubuntu 20.04 is shipped with 3.0.4, but the Lazarus project provides
DEB packages for new versions. On the other hand, we can finally
remove the workaround for gzopen() permissions, which was fixed in
3.1.1: https://gitlab.com/freepascal.org/fpc/source/-/issues/28917